[TestNG] Always map Pickles/TestCases to TestNG tests.
Remove the support for mapping the whole Cucumber test suite or each
Cucumber feature to TestNG tests.
