+++ testng/src/main/java/cucumber/api/testng/PickleEventWrapper.java
@@ -2,20 +2,12 @@ package cucumber.api.testng;
 
 import gherkin.events.PickleEvent;
 
-public class PickleEventWrapper {
-
-    private final PickleEvent pickleEvent;
-
-    PickleEventWrapper(PickleEvent pickleEvent) {
-        this.pickleEvent = pickleEvent;
-    }
-
-    public PickleEvent getPickleEvent() {
-        return pickleEvent;
-    }
-
-    @Override
-    public String toString() {
-        return "\"" + pickleEvent.pickle.getName() + "\"";
-    }
+/**
+ * The only purpose of this interface is to be able to provide a custom
+ * <pre>toString()</pre>, making TestNG reports look more descriptive.
+ *
+ * @see PickleEventWrapperImpl
+ */
+public interface PickleEventWrapper {
+    public PickleEvent getPickleEvent();
 }
