+++ testng/src/test/java/cucumber/api/testng/TestNGCucumberRunnerTest.java
@@ -9,9 +9,6 @@ import org.testng.Assert;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-import java.io.File;
-import java.io.FilenameFilter;
-import java.net.URL;
 import java.util.List;
 
 public class TestNGCucumberRunnerTest {
@@ -23,9 +20,13 @@ public class TestNGCucumberRunnerTest {
     }
 
     @Test(expectedExceptions = CucumberException.class)
-    public void runCukesStrict() throws Exception {
+    public void runCukesStrict() throws Throwable {
         testNGCucumberRunner = new TestNGCucumberRunner(RunCukesStrict.class);
-        testNGCucumberRunner.runCukes();
+
+        for (Object[] scenario : testNGCucumberRunner.provideScenarios()) {
+            PickleEventWrapper pickleEvent = (PickleEventWrapper) scenario[0];
+            testNGCucumberRunner.runScenario(pickleEvent.getPickleEvent());
+        }
     }
 
     @Test(expectedExceptions = CucumberException.class)
@@ -44,38 +45,13 @@ public class TestNGCucumberRunnerTest {
     @Test
     public void parse_error_propagated_to_testng_test_execution() throws Exception {
         testNGCucumberRunner = new ParseErrorCucumberRunner(RunCukesTest.class);
-        Object[][] features = testNGCucumberRunner.provideScenarios(); // provideScenarios() throws CucumberException
+        Object[][] scenarios = testNGCucumberRunner.provideScenarios(); // a CucumberException is caught
         try {
-            ((CucumberFeatureWrapper)features[0][0]).getCucumberFeature();
+            ((PickleEventWrapper)scenarios[0][0]).getPickleEvent();
             Assert.fail("CucumberException not thrown");
         } catch (CucumberException e) {
             Assert.assertEquals(e.getMessage(), "parse error");
         }
-    }
-
-    @Test
-    public void getFeatures() throws Exception {
-        List<CucumberFeature> features = testNGCucumberRunner.getFeatures();
-
-        int numberOfFeatures = getNumberOfFeatures();
-
-        Assert.assertEquals(features.size(), numberOfFeatures,
-                "Not all features associated with " + RunCukesTest.class.getSimpleName() + " were loaded. ");
-        Assert.assertTrue(!features.isEmpty(), "Feature files need to exist in the cucumber/runtime/testng/ folder for this test");
-    }
-
-    /**
-     * @return number of feature files in "cucumber/runtime/testng/" folder
-     */
-    private int getNumberOfFeatures() {
-        URL fileURL = this.getClass().getClassLoader().getResource("cucumber/runtime/testng/");
-        assert fileURL != null;
-        return new File(fileURL.getFile()).listFiles(new FilenameFilter() {
-            @Override
-            public boolean accept(File dir, String name) {
-                return name.endsWith(".feature");
-            }
-        }).length;
     }
 }
 
@@ -85,7 +61,7 @@ class ParseErrorCucumberRunner extends TestNGCucumberRunner {
     }
 
     @Override
-    public List<CucumberFeature> getFeatures() {
+    List<CucumberFeature> getFeatures() {
         throw new CucumberException("parse error");
     }
 }
